Added demo application source files
